+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - Removed "rescript legacy" subcommand in favor of separate "rescript-legacy" binary. https://github.com/rescript-lang/rescript/pull/7928
 - Add comparison fn for Error in Result.equal and compare. https://github.com/rescript-lang/rescript/pull/7933
+- Rewatch: `"type": "dev"` and `dev-dependencies` will be compiled by default for local packages. The `--dev` flag no longer does anything. https://github.com/rescript-lang/rescript/pull/7934
 
 #### :eyeglasses: Spec Compliance